chore(repo): restructure for seperation of concerns
