+++ src/config/gpt_config.py
@@ -1,0 +1,9 @@
+from dataclasses import dataclass
+
+@dataclass 
+class GPTConfig:
+    block_size: int = 1024
+    vocab_size: int = 50257
+    n_layer: int = 12
+    n_head: int = 12
+    n_embd: int = 768
